feat: merge day and week view
BREAKING CHANGE:
The day view is now the same as the week view. The following exported functions and interfaces have changed as follows:

`getDayView` -> was removed, use `getWeekView` with the `viewStart` and `viewEnd` options being set to the start and end of a single day.

`getDayViewHourGrid` -> was removed, its now part of the `getWeekView` function.

`DayViewHourSegment` -> was renamed to `WeekViewHourSegment`.

`DayViewHour` -> was renamed to `WeekViewHour`

`DayViewEvent` -> was renamed to `WeekViewTimeEvent`

`getWeekViewEventOffset` -> was removed from the public api

`DayView` -> was removed from the public api